+++ examples/apps/web-table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@2.4.0
+
 ## 0.1.8
 
 ### Patch Changes